Update `.gitignore` file
Fixes issue #50

This updates the `.gitignore` file so that it points to the directories as located in the root folder of this project.
I am not sure if anythihng is packaged in an egg in the first place, and whether the associated info file should be removed, but I kept it in there to be safe.